baby_object_detection/requirements.txt
ultralytics
streamlit
Pillow
c